Move components related to budget summary into their own subdirectory
