+++ src/app/App.js
@@ -9,7 +9,7 @@ import Environment from 'environment/Environment';
 import AppLayout from 'app/AppLayout';
 import AppRoutes from 'app/AppRoutes';
 import AuthenticationIndicator from 'AuthenticationIndicator';
-import Budget from 'Budget';
+import Budget from 'summary/Budget';
 
 export default class App extends React.Component {
     constructor(props) {
